LF-4768 Cleanup refactors
