+++ packages/webapp/src/components/IrrigationPrescription/types.ts
@@ -28,3 +28,45 @@ export type VriPrescriptionData = UriPrescriptionData & {
 export type IrrigationPrescriptionTableInfo = UriPrescriptionData & {
   id: number;
 };
+
+// TODO LF-4788: This is the backend-received data structure and belongs in the RTK Query slice types once that endpoint is implemented
+export type IrrigationPrescription = {
+  id: number;
+
+  location_id: string;
+  recommended_start_datetime: string; // ISO string
+
+  pivot: {
+    center: { lat: number; lng: number };
+    radius: number; // in meters
+  };
+
+  metadata: {
+    // metadata = external sources of information used to generate the irrigation prescription
+    weather_forecast?: {
+      temperature: number;
+      temperature_unit: string;
+      wind_speed: number;
+      wind_speed_unit: string;
+      cumulative_rainfall: number;
+      cumultative_rainfall_unit: string;
+      et_rate: number;
+      et_rate_unit: string;
+      weather_icon_code: string; // '02d', '50n', OpenWeatherMap icon code if available
+    };
+  };
+
+  estimated_time: number;
+  estimated_time_unit: string;
+
+  // TODO: confirm with product if we are indeed getting only URI or VRI data per prescription
+  prescription:
+    | { uriData: UriPrescriptionData; vriData?: never }
+    | {
+        vriData: {
+          zones: VriPrescriptionData[];
+          file_url: string;
+        };
+        uriData?: never;
+      };
+};
