+++ packages/webapp/src/containers/IrrigationPrescription/index.tsx
@@ -22,16 +22,17 @@ import { Main } from '../../components/Typography';
 import { DateInput, TimeInput } from '../../components/Inputs/DateTime';
 import PureIrrigationPrescription from '../../components/IrrigationPrescription';
 import type { CustomRouteComponentProps } from '../../types';
+import PageTitle from '../../components/PageTitle/v2';
+import Layout from '../../components/Layout';
+import layoutStyles from '../../components/Layout/layout.module.scss';
 import { Location } from '../../types';
+import type { IrrigationPrescription } from '../../components/IrrigationPrescription/types';
 import {
   mockField,
   mockUriData,
   mockVriZones,
   mockPivot,
 } from '../../stories/IrrigationPrescription/mockData';
-import PageTitle from '../../components/PageTitle/v2';
-import Layout from '../../components/Layout';
-import layoutStyles from '../../components/Layout/layout.module.scss';
 
 interface RouteParams {
   ip_pk: string;
@@ -44,60 +45,6 @@ const dateTimeLabelStyles = {
   fontSize: '16px',
 };
 
-/* ------------------------------ */
-
-export interface UriPrescriptionData {
-  soil_moisture_deficit: number;
-  application_depth: number;
-  application_depth_unit: string;
-}
-
-export type VriPrescriptionData = UriPrescriptionData & {
-  grid_points: { lat: number; lng: number }[];
-};
-
-export type IrrigationPrescription = {
-  id: number;
-
-  location_id: string;
-  recommended_start_datetime: string; // ISO string
-
-  pivot: {
-    center: { lat: number; lng: number };
-    radius: number; // in meters
-  };
-
-  metadata: {
-    // metadata = external sources of information used to generate the irrigation prescription
-    weather_forecast?: {
-      temperature: number;
-      temperature_unit: string;
-      wind_speed: number;
-      wind_speed_unit: string;
-      cumulative_rainfall: number;
-      cumultative_rainfall_unit: string;
-      et_rate: number;
-      et_rate_unit: string;
-      weather_icon_code: string; // '02d', '50n', OpenWeatherMap icon code if available
-    };
-  };
-
-  estimated_time: number;
-  estimated_time_unit: string;
-
-  // TODO: confirm with product if we are indeed getting only URI or VRI data per prescription
-  prescription:
-    | { uriData: UriPrescriptionData; vriData?: never }
-    | {
-        vriData: {
-          zones: VriPrescriptionData[];
-          file_url: string;
-        };
-        uriData?: never;
-      };
-};
-/* ------------------------------ */
-
 const IrrigationPrescription = ({ match }: IrrigationPrescriptionProps) => {
   const { t } = useTranslation();
 
@@ -109,11 +56,9 @@ const IrrigationPrescription = ({ match }: IrrigationPrescriptionProps) => {
   
   TODO LF-4788: Call the backend here to get the actual data for the given uuid 
   
-  Also handle case of no matching uuid (unknown record)
+  Also handle case of no matching uuid (unknown record) */
 
-  --------------------------------------*/
-
-  const commonProps = {
+  const commonMockData = {
     location_id: mockField.location_id,
     recommended_start_datetime: new Date().toISOString(),
     pivot: mockPivot,
@@ -139,14 +84,14 @@ const IrrigationPrescription = ({ match }: IrrigationPrescriptionProps) => {
   const irrigationPrescription: IrrigationPrescription =
     ip_pk === 'uri_pk'
       ? {
-          ...commonProps,
+          ...commonMockData,
           id: 1,
           prescription: {
             uriData: mockUriData,
           },
         }
       : {
-          ...commonProps,
+          ...commonMockData,
           id: 2,
           prescription: {
             vriData: {
@@ -159,11 +104,12 @@ const IrrigationPrescription = ({ match }: IrrigationPrescriptionProps) => {
   const fieldLocation: Location | undefined =
     useSelector(locationByIdSelector(irrigationPrescription?.location_id ?? '')) || mockField;
 
+  /* ------------------------------------- */
+
   const { prescription, pivot, recommended_start_datetime } = irrigationPrescription;
 
   const { uriData, vriData } = prescription;
 
-  /* ------------------------------------- */
   return (
     <Layout className={layoutStyles.paperContainer}>
       <PageTitle title={t('IRRIGATION_PRESCRIPTION.TITLE')} onGoBack={() => history.back()} />
