+++ packages/webapp/src/containers/IrrigationPrescription/styles.module.scss
@@ -13,7 +13,6 @@
  *  GNU General Public License for more details, see <https://www.gnu.org/licenses/>.
  */
 .irrigationPrescriptionContainer {
-  max-width: 100vw;
   display: flex;
   flex-direction: column;
   gap: 16px;
@@ -23,19 +22,13 @@
 
 .recommendedSchedule {
   display: flex;
-  padding: 16px 0px;
   flex-direction: column;
-  align-items: flex-start;
-  align-self: stretch;
-
   gap: 16px;
   padding: 16px;
 
+  background: var(--White);
   border-radius: 8px;
-  border: 1px solid var(--Colors-Primary-Primary-teal-50, #ebf5f4);
-  background: var(--White, #fff);
-
-  /* simple-box-shadow */
+  border: 1px solid var(--Colors-Primary-Primary-teal-50);
   box-shadow: 0px 1px 0px 0px rgba(0, 0, 0, 0.05);
 
   > div {
